Fixed appId type

+++ internal/domain/models/app.go
@@ -3,7 +3,7 @@ package models
 import "time"
 
 type App struct {
-	ID          int64
+	ID          int32
 	CreatedAt   time.Time
 	UpdatedAt   time.Time
 	Name        string
